requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
spotipy==2.19.0
python-dotenv==0.19.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
psutil==5.9.5